pyproject.toml
[tool.poetry]
name = "demo-project"
version = "0.1.0"
description = ""
authors = ["Keith Lagrange <punolagrange@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aws-cdk-lib = "2.153"
constructs = "^10.3.0"
pytest = "^8.3.2"
flake8 = "^7.1.1"
black = "^24.8.0"
compliant-cdk-lib = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
